go back to std::unordered_map
std::set just isn't worth it.
we get a slight win in memory usage (since key and value are not
duplicated), but that win is constant (x2). Whereas we kinda have a
loss in performance with std::set:
- we do one lookup before insertion, to avoid another needless
std::string allocation on the heap
- we then need to insert again if it wasn't present

this is due to the fact that unordered_set/set have read-only keys.
which makes sense, unless you the hash/equal_to function doesn't
change depending if the key_type changes. i.e. in our case, they are
identical if pointee is the same... But we can't prove that to the
compiler. So I guess it's just easier to make it const. Sad! But
understandable...